Added method for composing and sending message
added composeMessage(char* type,Contents* list, WiFiClient KPI) : receives a list of structs containing the type of content (e.g. transaction_id) , the content itself and a pointer to the next content. Type is need to find the right template that's going to be filled with contents ; the transmission of the message is done directly in the function itself.

added join(short trId, char nodeId[], WiFiClient KPI) : calls composeMessage after having built the content list