storing/renaming is now working
The store() function is now working as expected. Its purpose is to read the output file generated by the receive() function (which reads the server's responses) and find the corresponding transaction type and id , in order to create a copy of the original file whose name is composed by those informations , for both a future reading of the newly created file , and to leave space for another receive() call.
In the future it might be included in the receive() function itself , but for now ,given the unreliability and the unexpected data losses of the wifi shield , it's better to play safe.